Data Architecture/TenK_Google_Functions/ten-k-data-upload-placeholder-status-options-1/requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudevents==1.9.0
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
deprecation==2.1.0
docopt==0.6.2
executing==1.2.0
Flask==2.2.3
functions-framework==3.3.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
google-cloud-bigquery==3.*
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.1.0
jupyter_core==5.3.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.0
pandas==2.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.13
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.0.2
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
yarg==0.1.9
